added nav bar, reorganized components into diff files
